+++ derive_builder_core/src/builder.rs
@@ -12,6 +12,11 @@ use BuilderPattern;
 use DeprecationNotes;
 use Setter;
 
+const ALLOC_NOT_ENABLED_ERROR: &str = r#"`alloc` is disabled within 'derive_builder', consider one of the following:
+* enable feature `alloc` on 'dervie_builder' if a `global_allocator` is present
+* use a custom error `#[builder(build_fn(error = "path::to::Error"))]
+* disable the validation error `#[builder(build_fn(error(validation_error = false)))]"#;
+
 /// Builder, implementing `quote::ToTokens`.
 ///
 /// # Examples
@@ -129,6 +134,14 @@ pub struct Builder<'a> {
     ///
     /// This would be `false` in the case where an already-existing error is to be used.
     pub generate_error: bool,
+    /// Whether to include `ValidationError` in the generated enum. Necessary to avoid dependency
+    /// on `alloc::string`.
+    ///
+    /// This would be `false` when `build_fn.error.as_validation_error() == Some((false, _))`. This
+    /// has no effect when `generate_error` is `false`.
+    pub generate_validation_error: bool,
+    /// Indicator of `cfg!(not(any(feature = "alloc", feature = "std")))`, as a field for tests
+    pub no_alloc: bool,
     /// Whether this builder must derive `Clone`.
     ///
     /// This is true even for a builder using the `owned` pattern if there is a field whose setter
@@ -227,9 +240,39 @@ impl<'a> ToTokens for Builder<'a> {
                 ));
             }
 
-            if self.generate_error {
+            if self.no_alloc && self.generate_error && self.generate_validation_error {
+                let err = syn::Error::new_spanned(&self.ident, ALLOC_NOT_ENABLED_ERROR);
+                tokens.append_all(err.to_compile_error());
+            } else if self.generate_error {
                 let builder_error_ident = format_ident!("{}Error", builder_ident);
                 let builder_error_doc = format!("Error type for {}", builder_ident);
+
+                let validation_error = if self.generate_validation_error {
+                    quote!(
+                        /// Custom validation error
+                        ValidationError(#crate_root::export::core::string::String),
+                    )
+                } else {
+                    TokenStream::new()
+                };
+                let validation_from = if self.generate_validation_error {
+                    quote!(
+                        impl #crate_root::export::core::convert::From<#crate_root::export::core::string::String> for #builder_error_ident {
+                            fn from(s: #crate_root::export::core::string::String) -> Self {
+                                Self::ValidationError(s)
+                            }
+                        }
+                    )
+                } else {
+                    TokenStream::new()
+                };
+                let validation_display = if self.generate_validation_error {
+                    quote!(
+                        Self::ValidationError(ref error) => write!(f, "{}", error),
+                    )
+                } else {
+                    TokenStream::new()
+                };
 
                 tokens.append_all(quote!(
                     #[doc=#builder_error_doc]
@@ -238,8 +281,7 @@ impl<'a> ToTokens for Builder<'a> {
                     #builder_vis enum #builder_error_ident {
                         /// Uninitialized field
                         UninitializedField(&'static str),
-                        /// Custom validation error
-                        ValidationError(#crate_root::export::core::string::String),
+                        #validation_error
                     }
 
                     impl #crate_root::export::core::convert::From<#crate_root::UninitializedFieldError> for #builder_error_ident {
@@ -248,17 +290,13 @@ impl<'a> ToTokens for Builder<'a> {
                         }
                     }
 
-                    impl #crate_root::export::core::convert::From<#crate_root::export::core::string::String> for #builder_error_ident {
-                        fn from(s: #crate_root::export::core::string::String) -> Self {
-                            Self::ValidationError(s)
-                        }
-                    }
+                    #validation_from
 
                     impl #crate_root::export::core::fmt::Display for #builder_error_ident {
                         fn fmt(&self, f: &mut #crate_root::export::core::fmt::Formatter) -> #crate_root::export::core::fmt::Result {
                             match self {
                                 Self::UninitializedField(ref field) => write!(f, "`{}` must be initialized", field),
-                                Self::ValidationError(ref error) => write!(f, "{}", error),
+                                #validation_display
                             }
                         }
                     }
@@ -356,6 +394,8 @@ macro_rules! default_builder {
             field_initializers: vec![quote!(foo: ::db::export::core::default::Default::default(), )],
             functions: vec![quote!(fn bar() -> { unimplemented!() })],
             generate_error: true,
+            generate_validation_error: true,
+            no_alloc: false,
             must_derive_clone: true,
             doc_comment: None,
             deprecation_notes: DeprecationNotes::default(),
@@ -370,6 +410,43 @@ mod tests {
     use super::*;
     use proc_macro2::TokenStream;
     use syn::Ident;
+
+    fn add_simple_foo_builder(result: &mut TokenStream) {
+        #[cfg(not(feature = "clippy"))]
+        result.append_all(quote!(#[allow(clippy::all)]));
+
+        result.append_all(quote!(
+            #[derive(Clone)]
+            pub struct FooBuilder {
+                foo: u32,
+            }
+        ));
+
+        #[cfg(not(feature = "clippy"))]
+        result.append_all(quote!(#[allow(clippy::all)]));
+
+        result.append_all(quote!(
+            #[allow(dead_code)]
+            impl FooBuilder {
+                fn bar () -> {
+                    unimplemented!()
+                }
+
+                /// Create an empty builder, with all fields set to `None` or `PhantomData`.
+                fn create_empty() -> Self {
+                    Self {
+                        foo: ::db::export::core::default::Default::default(),
+                    }
+                }
+            }
+
+            impl ::db::export::core::default::Default for FooBuilder {
+                fn default() -> Self {
+                    Self::create_empty()
+                }
+            }
+        ));
+    }
 
     fn add_generated_error(result: &mut TokenStream) {
         result.append_all(quote!(
@@ -417,40 +494,7 @@ mod tests {
             {
                 let mut result = quote!();
 
-                #[cfg(not(feature = "clippy"))]
-                result.append_all(quote!(#[allow(clippy::all)]));
-
-                result.append_all(quote!(
-                    #[derive(Clone)]
-                    pub struct FooBuilder {
-                        foo: u32,
-                    }
-                ));
-
-                #[cfg(not(feature = "clippy"))]
-                result.append_all(quote!(#[allow(clippy::all)]));
-
-                result.append_all(quote!(
-                    #[allow(dead_code)]
-                    impl FooBuilder {
-                        fn bar () -> {
-                            unimplemented!()
-                        }
-
-                        /// Create an empty builder, with all fields set to `None` or `PhantomData`.
-                        fn create_empty() -> Self {
-                            Self {
-                                foo: ::db::export::core::default::Default::default(),
-                            }
-                        }
-                    }
-
-                    impl ::db::export::core::default::Default for FooBuilder {
-                        fn default() -> Self {
-                            Self::create_empty()
-                        }
-                    }
-                ));
+                add_simple_foo_builder(&mut result);
 
                 add_generated_error(&mut result);
 
@@ -751,6 +795,70 @@ mod tests {
                 ));
 
                 add_generated_error(&mut result);
+
+                result
+            }
+            .to_string()
+        );
+    }
+
+    #[test]
+    fn no_validation_error() {
+        let mut builder = default_builder!();
+        builder.generate_validation_error = false;
+
+        assert_eq!(
+            quote!(#builder).to_string(),
+            {
+                let mut result = quote!();
+
+                add_simple_foo_builder(&mut result);
+
+                result.append_all(quote!(
+                    #[doc="Error type for FooBuilder"]
+                    #[derive(Debug)]
+                    #[non_exhaustive]
+                    pub enum FooBuilderError {
+                        /// Uninitialized field
+                        UninitializedField(&'static str),
+                    }
+
+                    impl ::db::export::core::convert::From<::db::UninitializedFieldError> for FooBuilderError {
+                        fn from(s: ::db::UninitializedFieldError) -> Self {
+                            Self::UninitializedField(s.field_name())
+                        }
+                    }
+
+                    impl ::db::export::core::fmt::Display for FooBuilderError {
+                        fn fmt(&self, f: &mut ::db::export::core::fmt::Formatter) -> ::db::export::core::fmt::Result {
+                            match self {
+                                Self::UninitializedField(ref field) => write!(f, "`{}` must be initialized", field),
+                            }
+                        }
+                    }
+
+                    impl std::error::Error for FooBuilderError {}
+                ));
+
+                result
+            }
+                .to_string()
+        );
+    }
+
+    #[test]
+    fn no_alloc_bug_using_string() {
+        let mut builder = default_builder!();
+        builder.no_alloc = true;
+
+        assert_eq!(
+            quote!(#builder).to_string(),
+            {
+                let mut result = quote!();
+
+                add_simple_foo_builder(&mut result);
+
+                result.append_all(quote!(compile_error! { #ALLOC_NOT_ENABLED_ERROR }));
 
                 result
             }